Modularize all methods except those involving git
