+++ api/editor/addLabel.js
@@ -1,0 +1,60 @@
+// Require any modules here.
+
+// Quick versions of document.querySelector and document.querySelectorAll
+const { q, qA } = require('../../modules/queries.js');
+
+// This outer function allows the API method access to the API and paths objects.
+// If you don't need the paths object, you do not need to include it.
+// DO NOT add more parameters to this function.
+// Note that URIs inside either of these functions are relative to api.js, not this file.
+module.exports = (api, paths, extra) => {
+  // You can put variables your code needs to access between runs here.
+  const project = extra.project;
+  // hex to hsl based on https://gist.github.com/xenozauros/f6e185c8de2a04cdfecf
+  function hexToHSL(hex) {
+    var result = /^#?([a-f\d]{2})([a-f\d]{2})([a-f\d]{2})$/i.exec(hex);
+      r = parseInt(result[1], 16);
+      g = parseInt(result[2], 16);
+      b = parseInt(result[3], 16);
+      r /= 255, g /= 255, b /= 255;
+      var max = Math.max(r, g, b), min = Math.min(r, g, b);
+      var h, s, l = (max + min) / 2;
+      if(max == min){
+        h = s = 0; // achromatic
+      }else{
+        var d = max - min;
+        s = l > 0.5 ? d / (2 - max - min) : d / (max + min);
+        switch(max){
+          case r: h = (g - b) / d + (g < b ? 6 : 0); break;
+          case g: h = (b - r) / d + 2; break;
+          case b: h = (r - g) / d + 4; break;
+        }
+        h /= 6;
+      }
+    var HSL = new Object();
+    HSL['hue']=Math.round(h*360);
+    HSL['saturation']=Math.round(s*100);
+    HSL['lightness']=Math.round(l*100);
+    return HSL;
+  }
+
+  //This is the final function that will become part of the API.
+  // You MAY make it async.
+  // You MAY add parameters.
+  function returnFunction(name, color, description = '') {
+    const label = {
+      id: api.idFromPath(api.fileName()),
+      name: name,
+      description: '',
+      color: hexToHSL(color)
+    };
+
+    project.labels.push(label);
+
+    api.saveProject();
+    api.populateFiletree();
+    api.updateLabelCSS();
+  }
+
+  return returnFunction;
+};
