+++ api/editor/openItem.js
@@ -4,17 +4,23 @@
 const { q, qA } = require('../../modules/queries.js');
 
 // This outer function allows the API method access to the API and paths objects.
-// extra is an object with any extra variables you might need.
+// If you don't need the paths object, you do not need to include it.
 // DO NOT add more parameters to this function.
 // Note that URIs inside either of these functions are relative to api.js, not this file.
 module.exports = (api, paths, extra) => {
   // You can put variables your code needs to access between runs here.
+  const project = extra.project;
 
   //This is the final function that will become part of the API.
   // You MAY make it async.
   // You MAY add parameters.
-  function returnFunction() {
-
+  function returnFunction(id, editorIndex = 0) {
+    api.focusItem(id);
+    const file = api.flatten(project.index).find(i => api.idFromPath(i.path) === id);
+    api.openFile(api.idFromPath(file.path), file.name, editorIndex);
+    project.openFile = id;
+    api.saveProject();
+    return document.getElementById(id);
   }
 
   return returnFunction;
