+++ api/editor/suggestWords.js
@@ -9,12 +9,13 @@ const { q, qA } = require('../../modules/queries.js');
 // Note that URIs inside either of these functions are relative to api.js, not this file.
 module.exports = (api, paths, extra) => {
   // You can put variables your code needs to access between runs here.
+  const dictionary = extra.dictionary;
 
   //This is the final function that will become part of the API.
   // You MAY make it async.
   // You MAY add parameters.
-  function returnFunction() {
-
+  function returnFunction(w) {
+    return dictionary.suggest(w);
   }
 
   return returnFunction;
