+++ frontend/assets/js/editor.js
@@ -131,7 +131,7 @@ function moveItem(event, index, main = false) {
 
 /* Modals */
 function updateProjectDetails() {
-  api.updateDetails({
+  api.updateProjectDetails({
     title: document.getElementById('projectDetails__title').value,
     author: document.getElementById('projectDetails__author').value,
     synopsis: document.getElementById('projectDetails__synopsis').value,
@@ -202,8 +202,8 @@ function resetSprint() {
 }
 
 /* Labels */
-function createLabel() {
-  return api.createLabel(q('#createLabel__name').value, q('#createLabel__color').value, q('#createLabel__description').value);
+function addLabel() {
+  return api.addLabel(q('#addLabel__name').value, q('#addLabel__color').value, q('#addLabel__description').value);
 }
 
 /* Tabs */
