+++ frontend/editor.html
@@ -176,23 +176,23 @@
         <p><em>Note: You will have to manually update your Title Page and any other references to this information in your novel's content.</em></p>
         <a href="#" onclick="this.parentNode.classList.remove('visible');updateProjectDetails()" class="button">Save and Close</a>
       </div>
-      <div id="createLabel" class="modal">
+      <div id="addLabel" class="modal">
         <h3>Create Label</h3>
         <form>
           <label>
             Name
-            <input placeholder="Something short but memorable..." type="text" id="createLabel__name" required>
+            <input placeholder="Something short but memorable..." type="text" id="addLabel__name" required>
           </label>
           <label>
             Description (optional)
-            <textarea id="createLabel__description" placeholder="What this label means to you..."></textarea>
+            <textarea id="addLabel__description" placeholder="What this label means to you..."></textarea>
           </label>
           <label>
             Color
-            <input type="color" id="createLabel__color" value="#ff0000"></textarea>
+            <input type="color" id="addLabel__color" value="#ff0000"></textarea>
           </label>
         </form>
-        <a href="#" onclick="this.parentNode.classList.remove('visible');createLabel()" class="button">Save and Close</a>
+        <a href="#" onclick="this.parentNode.classList.remove('visible');addLabel()" class="button">Save and Close</a>
         <a href="#" onclick="this.parentNode.classList.remove('visible');" class="button">Cancel</a>
       </div>
       <div id="projectGoalComplete" class="modal">
